Another update for the i3-git pacman41 comments.
Just set the branch in the source field, removes the need for choosing a
branch later on.

Note, though, that when you check the version of i3 with `i3 --version',
it will not return `next' anywhere, but `makepkg'. Pacman makes its own
working branch called makepkg just for this. (I guess you could leave in
the original stuff and put a note in for it...but the git version string
is totally fine, so it shouldn't matter too much.)